+++ CHANGELOG.md
@@ -5,6 +5,13 @@ Note: Prior to version 0.3.4.0, this library was named
 `small-bytearray-builder` is now just a compatibility shim
 to ease the migration process.
 
+## 0.3.16.0 -- 2024-01-29
+
+* Add `wordPaddedDec3`.
+* Add `Data.Bytes.Builder.Avro`.
+* Add `word16LEB128`.
+* Stop accepting versions of text lower than 2.0.
+
 ## 0.3.15.0 -- 2024-01-05
 
 * Add `Data.Bytes.Builder.Unsafe.pasteUtf8TextJson#` for users who need